% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materials.R
\name{prep_deployment_questions}
\alias{prep_deployment_questions}
\title{Prepare Deployment Questions}
\usage{
prep_deployment_questions(deployment_id, x = NULL)
}
\arguments{
\item{deployment_id}{Deployment ID.}

\item{x}{Table with questions or \code{NULL}.}
}
\description{
Prepare Deployment Questions
}
