% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materials.R
\name{prep_model_metadata}
\alias{prep_model_metadata}
\title{Prepare Model Metadata}
\usage{
prep_model_metadata(
  x,
  model_id,
  user_id,
  material_settings = "[]",
  con = NULL,
  update = FALSE,
  ...
)
}
\arguments{
\item{x}{ODMAP table.}

\item{model_id}{Model ID.}

\item{user_id}{User ID.}

\item{material_settings}{Material settings.}

\item{con}{Database connection (use \code{NULL} to avoid inserting new record).}

\item{update}{Logical. Is this a new entry or an update to an existing one.}

\item{...}{Arguments passed to the write function.}
}
\description{
Prepare Model Metadata
}
