% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materials.R
\name{prep_model_summary}
\alias{prep_model_summary}
\title{Prepare Model Summary}
\usage{
prep_model_summary(
  x,
  species,
  model_id,
  species_id,
  user_id,
  material_settings = "[]",
  con = NULL
)
}
\arguments{
\item{x}{Table with multi species model summary.}

\item{species}{Reference species table.}

\item{model_id}{Model ID.}

\item{species_id}{Species ID.}

\item{user_id}{User ID.}

\item{material_settings}{Material settings.}

\item{con}{Database connection (use \code{NULL} to avoid inserting new record).}
}
\description{
Prepare Model Summary
}
