% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materials.R
\name{prep_predictor_metadata}
\alias{prep_predictor_metadata}
\title{Prepare Predictor Metadata}
\usage{
prep_predictor_metadata(
  x,
  model_id,
  user_id,
  material_settings = "[]",
  con = NULL
)
}
\arguments{
\item{x}{Predictor metadata table.}

\item{model_id}{Model ID.}

\item{user_id}{User ID.}

\item{material_settings}{Material settings.}

\item{con}{Database connection (use \code{NULL} to avoid inserting new record).}
}
\description{
Prepare Predictor Metadata
}
