% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materials.R
\name{prep_predictor_raster}
\alias{prep_predictor_raster}
\title{Prepare Predictor Raster}
\usage{
prep_predictor_raster(
  x,
  resolution = 5,
  model_id,
  user_id,
  material_settings = "[]",
  con = NULL,
  update = FALSE,
  ...
)
}
\arguments{
\item{x}{Raster file.}

\item{resolution}{Raster resampling resolution}

\item{model_id}{Model ID.}

\item{user_id}{User ID.}

\item{material_settings}{Material settings.}

\item{con}{Database connection (use \code{NULL} to avoid inserting new record).}

\item{update}{Logical. Is this a new entry or an update to an existing one.}

\item{...}{Arguments passed to the write function.}
}
\description{
Prepare Predictor Raster
}
