% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materials.R
\name{prep_spatial_prediction}
\alias{prep_spatial_prediction}
\title{Prepare Spatial Prediction}
\usage{
prep_spatial_prediction(
  x,
  model_id,
  species_id,
  user_id,
  material_settings = "[]",
  con = NULL,
  update = FALSE,
  ...
)
}
\arguments{
\item{x}{Raster file.}

\item{model_id}{Model ID.}

\item{species_id}{Species ID.}

\item{user_id}{User ID.}

\item{material_settings}{Material settings.}

\item{con}{Database connection (use \code{NULL} to avoid inserting new record).}

\item{update}{Logical. Is this a new entry or an update to an existing one.}

\item{...}{Arguments passed to the write function.}
}
\description{
Prepare Spatial Prediction
}
