% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdmEvalToolUp-package.R
\docType{package}
\name{sdmEvalToolUp-package}
\alias{sdmEvalToolUp}
\alias{sdmEvalToolUp-package}
\title{sdmEvalToolUp: SDM Evaluation Tool Shiny Upload}
\description{
The goal of sdmEvalToolUp is to implement Shiny functionality according to the SDM Evaluation Tool requirements for uploading materials.
}
\author{
\strong{Maintainer}: Peter Solymos \email{psolymos@gmail.com} (\href{https://orcid.org/0000-0001-7337-1740}{ORCID})

Authors:
\itemize{
  \item Steffi LaZerte \email{sel@steffilazerte.ca} (\href{https://orcid.org/0000-0002-7690-8360}{ORCID})
}

}
\keyword{internal}
