% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materials.R
\name{write_file_and_add_material_entry}
\alias{write_file_and_add_material_entry}
\title{Write File and Insert/Update Material DB Entry}
\usage{
write_file_and_add_material_entry(
  x,
  model_id,
  species_id,
  component_id,
  user_id,
  deployment_id = NULL,
  material_settings = NULL,
  con = NULL,
  update = FALSE,
  ...
)
}
\arguments{
\item{x}{Object.}

\item{model_id}{Model ID.}

\item{species_id}{Species ID.}

\item{component_id}{Component ID.}

\item{user_id}{User ID.}

\item{deployment_id}{Deployment ID.}

\item{material_settings}{Material settings.}

\item{con}{Database connection (use \code{NULL} to avoid inserting a record).}

\item{update}{Logical. Is this a new entry (\code{FALSE}, default)
or an update (\code{TRUE}) to an existing material entry.}

\item{...}{Arguments passed to the write function.}
}
\description{
Write File and Insert/Update Material DB Entry
}
