% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materials.R
\name{prep_deployment_settings}
\alias{prep_deployment_settings}
\title{Prepare Deployment Settings}
\usage{
prep_deployment_settings(deployment_id, x)
}
\arguments{
\item{deployment_id}{Deployment ID.}

\item{x}{Lits with deployment.}
}
\description{
Prepare Deployment Settings
}
