% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materials.R
\name{prep_deployment_subunits}
\alias{prep_deployment_subunits}
\title{Prepare Deployment Subunits}
\usage{
prep_deployment_subunits(
  deployment_id,
  x = NULL,
  reference,
  n = c(20, 20),
  square = TRUE
)
}
\arguments{
\item{deployment_id}{Deployment ID.}

\item{x}{A spatial file with subunits or \code{NULL}.}

\item{reference}{Reference raster.}

\item{n}{Number of grid cells in x and y direction.}

\item{square}{Logical. If `FALSE``, create hexagonal grid.}
}
\description{
Prepare Deployment Subunits
}
