% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materials.R
\name{prep_observations}
\alias{prep_observations}
\title{Prepare Observations}
\usage{
prep_observations(
  x,
  species,
  model_id,
  species_id,
  user_id,
  material_settings = "[]",
  con = NULL
)

prep_observations(
  x,
  species,
  model_id,
  species_id,
  user_id,
  material_settings = "[]",
  con = NULL
)
}
\arguments{
\item{x}{Table with multi species observations.}

\item{species}{Reference species table.}

\item{model_id}{Model ID.}

\item{species_id}{Species ID.}

\item{user_id}{User ID.}

\item{material_settings}{Material settings.}

\item{con}{Database connection (use \code{NULL} to avoid inserting new record).}
}
\description{
Prepare Observations

Prepare Observations
}
